+++ Readme/Changelog.md
@@ -4,6 +4,14 @@
 
 > Unreleased changes in [`master`](https://github.com/ReSwift/ReSwift) scheduled to be integrated into the next release.
 
+**API Changes:**
+
+- `SwiftFlowInit` action has been renamed to `ReSwiftInit` - @vfn
+
+**Other:**
+
+- Update Documentation. Sample reducer code was slightly outdated - @vfn
+
 ## [0.2.4](https://github.com/ReSwift/ReSwift/releases/tag/0.2.4)
 
 *Released: 01/23/2015*